requirements.txt
streamlit
watchdog
contextily
geopandas
plotly
